refactor(cli): extract helpers from cli.go below LOC ceiling
Move config/badge/printers/watch/version/help/completion out of
cli.go into cmd_*.go and version.go; lower architecture maxLOC to 490.
