+++ internal/cli/cli.go
@@ -2,22 +2,18 @@ package cli
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
 	"io"
 	"os"
-	"os/signal"
 	"strings"
-	"syscall"
 	"time"
 
 	"go.klarlabs.de/coverctl/internal/application"
 	"go.klarlabs.de/coverctl/internal/domain"
 	"go.klarlabs.de/coverctl/internal/infrastructure/annotations"
 	"go.klarlabs.de/coverctl/internal/infrastructure/autodetect"
-	"go.klarlabs.de/coverctl/internal/infrastructure/badge"
 	"go.klarlabs.de/coverctl/internal/infrastructure/bitbucket"
 	"go.klarlabs.de/coverctl/internal/infrastructure/config"
 	"go.klarlabs.de/coverctl/internal/infrastructure/diff"
@@ -28,10 +24,8 @@ import (
 	"go.klarlabs.de/coverctl/internal/infrastructure/report"
 	"go.klarlabs.de/coverctl/internal/infrastructure/resolver"
 	"go.klarlabs.de/coverctl/internal/infrastructure/runners"
-	"go.klarlabs.de/coverctl/internal/infrastructure/watcher"
 	"go.klarlabs.de/coverctl/internal/infrastructure/wizard"
 	"go.klarlabs.de/coverctl/internal/mcp"
-	"go.klarlabs.de/coverctl/internal/pathutil"
 )
 
 type Service interface {
@@ -394,36 +388,6 @@ func (t *testArgsList) Set(value string) error {
 	return nil
 }
 
-func writeConfigFile(path string, cfg application.Config, stdout io.Writer, force bool) error {
-	if path == "-" {
-		return config.Write(stdout, cfg)
-	}
-	if !force {
-		if _, err := os.Stat(path); err == nil {
-			return fmt.Errorf("config %s already exists", path)
-		}
-	}
-	cleanPath, err := pathutil.ValidatePath(path)
-	if err != nil {
-		return fmt.Errorf("invalid path: %w", err)
-	}
-	file, err := os.Create(cleanPath) // #nosec G304 - path is validated above
-	if err != nil {
-		return err
-	}
-	defer func() { _ = file.Close() }()
-	return config.Write(file, cfg)
-}
-
-// validateConfig checks if the config file is valid without running tests
-func validateConfig(path string) error {
-	_, err := config.Loader{}.Load(path)
-	if err != nil {
-		return fmt.Errorf("invalid config: %w", err)
-	}
-	return nil
-}
-
 func usage(w io.Writer) {
 	fmt.Fprintf(w, `coverctl - Domain-driven coverage enforcement for any language
 
@@ -465,29 +429,6 @@ Run 'coverctl help <command>' for more information on a command.
 `, Version)
 }
 
-func writeBadgeFile(path string, percent float64, label, style string) error {
-	cleanPath, err := pathutil.ValidatePath(path)
-	if err != nil {
-		return fmt.Errorf("invalid path: %w", err)
-	}
-	file, err := os.Create(cleanPath) // #nosec G304 - path is validated above
-	if err != nil {
-		return err
-	}
-	defer func() { _ = file.Close() }()
-
-	badgeStyle := badge.StyleFlat
-	if style == "flat-square" {
-		badgeStyle = badge.StyleFlatSquare
-	}
-
-	return badge.Generate(file, badge.Options{
-		Label:   label,
-		Percent: percent,
-		Style:   badgeStyle,
-	})
-}
-
 // exitCodeWithCI outputs errors in GitHub Actions annotation format when CI mode is enabled
 func exitCodeWithCI(err error, code int, stderr io.Writer, global GlobalOptions) int {
 	if err == nil {
@@ -524,874 +465,3 @@ func remediationHintForError(err error) string {
 	}
 	return ""
 }
-
-func printIgnoreInfo(cfg application.Config, domains []domain.Domain, w io.Writer) {
-	fmt.Fprintln(w, "Configured exclude patterns:")
-	if len(cfg.Exclude) == 0 {
-		fmt.Fprintln(w, "  (none yet). Add patterns such as `internal/generated/*` to ignore generated proto domains.")
-	} else {
-		for _, pattern := range cfg.Exclude {
-			fmt.Fprintf(w, "  - %s\n", pattern)
-		}
-	}
-	fmt.Fprintln(w, "\nDomains tracked by the policy:")
-	for _, d := range domains {
-		fmt.Fprintf(w, "  - %s (matches: %s)\n", d.Name, strings.Join(d.Match, ", "))
-	}
-	fmt.Fprintln(w, "\nUse `exclude:` entries in `.coverctl.yaml` to skip generated folders (e.g., proto outputs) before running `coverctl check`.")
-}
-
-func printTrendResult(result application.TrendResult, w io.Writer) {
-	trendSymbol := "→"
-	switch result.Trend.Direction {
-	case domain.TrendUp:
-		trendSymbol = "↑"
-	case domain.TrendDown:
-		trendSymbol = "↓"
-	}
-
-	fmt.Fprintf(w, "Coverage Trend: %.1f%% %s %.1f%% (%+.1f%%)\n",
-		result.Previous, trendSymbol, result.Current, result.Trend.Delta)
-	fmt.Fprintln(w, "\nDomain Trends:")
-	for name, trend := range result.ByDomain {
-		symbol := "→"
-		switch trend.Direction {
-		case domain.TrendUp:
-			symbol = "↑"
-		case domain.TrendDown:
-			symbol = "↓"
-		}
-		fmt.Fprintf(w, "  %s: %s %+.1f%%\n", name, symbol, trend.Delta)
-	}
-	fmt.Fprintf(w, "\nHistory: %d entries\n", len(result.Entries))
-}
-
-func printSuggestResult(result application.SuggestResult, w io.Writer) {
-	fmt.Fprintln(w, "Threshold Suggestions:")
-	fmt.Fprintln(w, "")
-	fmt.Fprintf(w, "%-20s %10s %10s %12s  %s\n", "DOMAIN", "CURRENT", "MIN", "SUGGESTED", "REASON")
-	fmt.Fprintf(w, "%-20s %10s %10s %12s  %s\n", "------", "-------", "---", "---------", "------")
-	for _, s := range result.Suggestions {
-		change := ""
-		if s.SuggestedMin > s.CurrentMin {
-			change = "↑"
-		} else if s.SuggestedMin < s.CurrentMin {
-			change = "↓"
-		}
-		fmt.Fprintf(w, "%-20s %9.1f%% %9.1f%% %10.1f%% %s  %s\n",
-			s.Domain, s.CurrentPercent, s.CurrentMin, s.SuggestedMin, change, s.Reason)
-	}
-}
-
-func printDebtResult(result application.DebtResult, w io.Writer, format application.OutputFormat) {
-	if format == application.OutputJSON {
-		enc := json.NewEncoder(w)
-		enc.SetIndent("", "  ")
-		_ = enc.Encode(result)
-		return
-	}
-
-	// Text output
-	if len(result.Items) == 0 {
-		fmt.Fprintln(w, "No coverage debt found - all targets are met!")
-		fmt.Fprintf(w, "Health Score: %.1f%%\n", result.HealthScore)
-		return
-	}
-
-	fmt.Fprintln(w, "Coverage Debt Report")
-	fmt.Fprintln(w, "====================")
-	fmt.Fprintln(w, "")
-	fmt.Fprintf(w, "%-8s %-30s %10s %10s %10s %8s\n", "TYPE", "NAME", "CURRENT", "REQUIRED", "SHORTFALL", "LINES")
-	fmt.Fprintf(w, "%-8s %-30s %10s %10s %10s %8s\n", "----", "----", "-------", "--------", "---------", "-----")
-
-	for _, item := range result.Items {
-		name := item.Name
-		if len(name) > 30 {
-			name = "..." + name[len(name)-27:]
-		}
-		fmt.Fprintf(w, "%-8s %-30s %9.1f%% %9.1f%% %9.1f%% %8d\n",
-			item.Type, name, item.Current, item.Required, item.Shortfall, item.Lines)
-	}
-
-	fmt.Fprintln(w, "")
-	fmt.Fprintf(w, "Total Debt: %.1f%% shortfall across %d items\n", result.TotalDebt, len(result.Items))
-	fmt.Fprintf(w, "Estimated Lines Needing Tests: %d\n", result.TotalLines)
-	fmt.Fprintf(w, "Health Score: %.1f%%\n", result.HealthScore)
-}
-
-func printCompareResult(result application.CompareResult, w io.Writer, format application.OutputFormat) {
-	if format == application.OutputJSON {
-		enc := json.NewEncoder(w)
-		enc.SetIndent("", "  ")
-		_ = enc.Encode(result)
-		return
-	}
-
-	// Text output
-	fmt.Fprintln(w, "Coverage Comparison")
-	fmt.Fprintln(w, "===================")
-	fmt.Fprintln(w, "")
-
-	// Overall summary
-	deltaSign := "+"
-	if result.Delta < 0 {
-		deltaSign = ""
-	}
-	fmt.Fprintf(w, "Overall: %.1f%% → %.1f%% (%s%.1f%%)\n", result.BaseOverall, result.HeadOverall, deltaSign, result.Delta)
-	fmt.Fprintln(w, "")
-
-	// Domain deltas if available
-	if len(result.DomainDeltas) > 0 {
-		fmt.Fprintln(w, "Domain Changes:")
-		for domain, delta := range result.DomainDeltas {
-			sign := "+"
-			if delta < 0 {
-				sign = ""
-			}
-			if delta > 0.1 || delta < -0.1 {
-				fmt.Fprintf(w, "  %-20s %s%.1f%%\n", domain, sign, delta)
-			}
-		}
-		fmt.Fprintln(w, "")
-	}
-
-	// Improved files
-	if len(result.Improved) > 0 {
-		fmt.Fprintf(w, "Improved Files (%d):\n", len(result.Improved))
-		for i, f := range result.Improved {
-			if i >= 10 {
-				fmt.Fprintf(w, "  ... and %d more\n", len(result.Improved)-10)
-				break
-			}
-			fmt.Fprintf(w, "  %-50s %.1f%% → %.1f%% (+%.1f%%)\n", truncateLeft(f.File, 50), f.BasePct, f.HeadPct, f.Delta)
-		}
-		fmt.Fprintln(w, "")
-	}
-
-	// Regressed files
-	if len(result.Regressed) > 0 {
-		fmt.Fprintf(w, "Regressed Files (%d):\n", len(result.Regressed))
-		for i, f := range result.Regressed {
-			if i >= 10 {
-				fmt.Fprintf(w, "  ... and %d more\n", len(result.Regressed)-10)
-				break
-			}
-			fmt.Fprintf(w, "  %-50s %.1f%% → %.1f%% (%.1f%%)\n", truncateLeft(f.File, 50), f.BasePct, f.HeadPct, f.Delta)
-		}
-		fmt.Fprintln(w, "")
-	}
-
-	// Summary
-	fmt.Fprintf(w, "Summary: %d improved, %d regressed, %d unchanged\n", len(result.Improved), len(result.Regressed), result.Unchanged)
-}
-
-func truncateLeft(s string, maxLen int) string {
-	if len(s) <= maxLen {
-		return s
-	}
-	return "..." + s[len(s)-(maxLen-3):]
-}
-
-func runWatch(ctx context.Context, stdout, stderr io.Writer, svc Service, configPath, profile string, domains []string, global GlobalOptions, buildFlags application.BuildFlags) int {
-	// Create watcher
-	w, err := watcher.New(watcher.WithDebounce(500 * time.Millisecond))
-	if err != nil {
-		if global.CI {
-			fmt.Fprintf(stderr, "::error::failed to create watcher: %v\n", err)
-		} else {
-			fmt.Fprintf(stderr, "failed to create watcher: %v\n", err)
-		}
-		return 3
-	}
-	defer func() { _ = w.Close() }()
-
-	// Handle Ctrl+C gracefully
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	sigCh := make(chan os.Signal, 1)
-	signal.Notify(sigCh, os.Interrupt, syscall.SIGTERM)
-	go func() {
-		<-sigCh
-		if !global.IsQuiet() {
-			fmt.Fprintln(stdout, "\nStopping watch mode...")
-		}
-		cancel()
-	}()
-
-	if !global.IsQuiet() {
-		fmt.Fprintln(stdout, "Watching for file changes... (Ctrl+C to stop)")
-		fmt.Fprintln(stdout, "")
-	}
-
-	callback := func(runNumber int, runErr error) {
-		if !global.IsQuiet() {
-			fmt.Fprintf(stdout, "\n--- Run #%d at %s ---\n", runNumber, time.Now().Format("15:04:05"))
-		}
-		if runErr != nil {
-			if global.CI {
-				fmt.Fprintf(stderr, "::error::Coverage run failed: %v\n", runErr)
-			} else {
-				fmt.Fprintf(stderr, "Coverage run failed: %v\n", runErr)
-			}
-		} else if !global.IsQuiet() {
-			fmt.Fprintln(stdout, "Coverage run completed successfully")
-		}
-	}
-
-	opts := application.WatchOptions{
-		ConfigPath: configPath,
-		Profile:    profile,
-		Domains:    domains,
-		BuildFlags: buildFlags,
-	}
-
-	if err := svc.Watch(ctx, opts, w, callback); err != nil {
-		if ctx.Err() == context.Canceled {
-			return 0 // Normal exit on Ctrl+C
-		}
-		if global.CI {
-			fmt.Fprintf(stderr, "::error::watch error: %v\n", err)
-		} else {
-			fmt.Fprintf(stderr, "watch error: %v\n", err)
-		}
-		return 3
-	}
-	return 0
-}
-
-func printVersion(w io.Writer) {
-	fmt.Fprintf(w, "coverctl version %s\n", Version)
-	if Commit != "unknown" {
-		fmt.Fprintf(w, "  commit: %s\n", Commit)
-	}
-	if Date != "unknown" {
-		fmt.Fprintf(w, "  built:  %s\n", Date)
-	}
-}
-
-var commandHelpText = map[string]string{
-	"check": `coverctl check - Run coverage and enforce policy
-
-Usage:
-  coverctl check [flags]
-
-Aliases:
-  c
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile output path (default ".cover/coverage.out")
-      --from-profile     Use existing coverage profile instead of running tests
-  -d, --domain string    Filter to specific domain (repeatable)
-  -o, --output string    Output format: text|json|html|brief (default "text")
-                         Use 'brief' for single-line LLM/agent-optimized output
-      --show-delta       Show coverage change from previous run
-      --history string   History file path for delta display
-      --fail-under N     Fail if overall coverage is below N percent
-      --ratchet          Fail if coverage decreases from previous recorded value
-      --validate         Validate config file without running tests
-
-Build/Test Flags:
-      --tags string      Build tags (e.g., integration,e2e)
-      --race             Enable race detector
-      --short            Skip long-running tests
-  -v                     Verbose test output
-      --run string       Run only tests matching pattern
-      --timeout string   Test timeout forwarded to runner (e.g., 10m, 1h)
-      --max-runtime string  Hard ceiling on total runtime (default "15m"; 0 disables)
-      --test-arg string  Additional argument passed to go test (repeatable)
-
-Examples:
-  coverctl check
-  coverctl check -c custom.yaml
-  coverctl check --fail-under 80
-  coverctl check --ratchet
-  coverctl check --validate
-  coverctl check --from-profile --profile coverage.out
-  coverctl check --tags integration
-  coverctl check --race --timeout 30m
-  coverctl c -d core -d api`,
-
-	"run": `coverctl run - Run coverage only, produce artifacts
-
-Usage:
-  coverctl run [flags]
-
-Aliases:
-  r
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile output path (default ".cover/coverage.out")
-  -d, --domain string    Filter to specific domain (repeatable)
-
-Build/Test Flags:
-      --tags string      Build tags (e.g., integration,e2e)
-      --race             Enable race detector
-      --short            Skip long-running tests
-  -v                     Verbose test output
-      --run string       Run only tests matching pattern
-      --timeout string   Test timeout forwarded to runner (e.g., 10m, 1h)
-      --max-runtime string  Hard ceiling on total runtime (default "15m"; 0 disables)
-      --test-arg string  Additional argument passed to go test (repeatable)
-
-Examples:
-  coverctl run
-  coverctl run --tags integration
-  coverctl run --race -v
-  coverctl r -p coverage.out`,
-
-	"watch": `coverctl watch - Watch for file changes and re-run coverage
-
-Usage:
-  coverctl watch [flags]
-
-Aliases:
-  w
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile output path (default ".cover/coverage.out")
-  -d, --domain string    Filter to specific domain (repeatable)
-
-Build/Test Flags:
-      --tags string      Build tags (e.g., integration,e2e)
-      --race             Enable race detector
-      --short            Skip long-running tests
-  -v                     Verbose test output
-      --run string       Run only tests matching pattern
-      --timeout string   Test timeout forwarded to runner (e.g., 10m, 1h)
-      --max-runtime string  Hard ceiling on total runtime (default "15m"; 0 disables)
-      --test-arg string  Additional argument passed to go test (repeatable)
-
-Examples:
-  coverctl watch
-  coverctl watch --tags integration
-  coverctl w -d core`,
-
-	"init": `coverctl init - Interactive setup wizard
-
-Usage:
-  coverctl init [flags]
-
-Aliases:
-  i
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -f, --force            Overwrite existing config file
-      --no-interactive   Skip the interactive init wizard
-
-Examples:
-  coverctl init
-  coverctl i -f`,
-
-	"detect": `coverctl detect - Autodetect domains and write config
-
-Usage:
-  coverctl detect [flags]
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -f, --force            Overwrite config if it exists
-      --dry-run          Preview config without writing
-
-Examples:
-  coverctl detect
-  coverctl detect --dry-run
-  coverctl detect -f`,
-
-	"report": `coverctl report - Analyze an existing profile
-
-Usage:
-  coverctl report [flags]
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile path (default ".cover/coverage.out")
-  -d, --domain string    Filter to specific domain (repeatable)
-  -o, --output string    Output format: text|json|html|brief (default "text")
-                         Use 'brief' for single-line LLM/agent-optimized output
-      --show-delta       Show coverage change from previous run
-      --history string   History file path for delta display
-      --uncovered        Show only files with 0% coverage
-      --diff <ref>       Show coverage for files changed since git ref
-      --merge <file>     Merge additional coverage profile (repeatable)
-
-Examples:
-  coverctl report
-  coverctl report -p custom.out -o json
-  coverctl report -o html > coverage.html
-  coverctl report --uncovered
-  coverctl report --diff main
-  coverctl report --merge integration.out --merge e2e.out`,
-
-	"badge": `coverctl badge - Generate an SVG coverage badge
-
-Usage:
-  coverctl badge [flags]
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile path (default ".cover/coverage.out")
-  -o, --output string    Output file path (default "coverage.svg")
-      --label string     Badge label text (default "coverage")
-      --style string     Badge style: flat|flat-square (default "flat")
-
-Examples:
-  coverctl badge
-  coverctl badge -o badge.svg --style flat-square`,
-
-	"trend": `coverctl trend - Show coverage trends over time
-
-Usage:
-  coverctl trend [flags]
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile path (default ".cover/coverage.out")
-      --history string   History file path (default ".cover/history.json")
-  -o, --output string    Output format: text|json|html|brief (default "text")
-
-Examples:
-  coverctl trend
-  coverctl trend -o json`,
-
-	"record": `coverctl record - Record current coverage to history
-
-Usage:
-  coverctl record [flags]
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile path (default ".cover/coverage.out")
-      --history string   History file path (default ".cover/history.json")
-      --commit string    Git commit SHA (optional)
-      --branch string    Git branch name (optional)
-      --run              Run coverage before recording history
-  -l, --language string  Override language detection (go, python, nodejs, rust, java)
-  -d, --domain string    Filter to specific domain (repeatable)
-      --tags string      Build tags (e.g., integration,e2e)
-      --race             Enable race detector
-      --short            Skip long-running tests
-  -v                  Verbose test output
-      --test-run string  Run only tests matching pattern
-      --timeout string   Test timeout forwarded to runner (e.g., 10m, 1h)
-      --max-runtime string  Hard ceiling on total runtime (default "15m"; 0 disables)
-      --test-arg string  Additional argument passed to go test (repeatable)
-
-Examples:
-  coverctl record
-  coverctl record --commit abc123 --branch main
-  coverctl record --run --tags integration`,
-
-	"suggest": `coverctl suggest - Suggest optimal coverage thresholds
-
-Usage:
-  coverctl suggest [flags]
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile path (default ".cover/coverage.out")
-      --strategy string  Suggestion strategy: current|aggressive|conservative (default "current")
-      --apply            Update config with suggested thresholds
-  -f, --force            Overwrite config if it exists
-
-Examples:
-  coverctl suggest
-  coverctl suggest --strategy aggressive --apply`,
-
-	"debt": `coverctl debt - Show coverage debt report
-
-Usage:
-  coverctl debt [flags]
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile path (default ".cover/coverage.out")
-  -o, --output string    Output format: text|json|brief (default "text")
-
-Examples:
-  coverctl debt
-  coverctl debt -o json`,
-
-	"ignore": `coverctl ignore - Show configured excludes and ignore advice
-
-Usage:
-  coverctl ignore [flags]
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-
-Examples:
-  coverctl ignore`,
-
-	"compare": `coverctl compare - Compare coverage between two profiles
-
-Usage:
-  coverctl compare [flags]
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -b, --base string      Base coverage profile (required)
-  -H, --head string      Head coverage profile (default ".cover/coverage.out")
-  -o, --output string    Output format: text|json|brief (default "text")
-
-Examples:
-  coverctl compare --base main.out --head feature.out
-  coverctl compare -b main.out -o json`,
-
-	"pr-comment": `coverctl pr-comment - Post coverage report as PR/MR comment
-
-Supports GitHub, GitLab, and Bitbucket. Provider is auto-detected from
-environment variables or can be specified with --provider.
-
-Usage:
-  coverctl pr-comment [flags]
-
-Flags:
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile path (default ".cover/coverage.out")
-      --base string      Base coverage profile for comparison (optional)
-      --pr int           Pull request/MR number (required, auto-detected on GitLab/Bitbucket)
-      --owner string     Repository owner/namespace (auto-detected from env)
-      --repo string      Repository name (auto-detected from env)
-      --provider string  Git provider: github, gitlab, bitbucket, or auto (default "auto")
-      --update           Update existing comment instead of creating new (default true)
-      --dry-run          Generate comment without posting
-
-Environment Variables:
-  GitHub:
-    GITHUB_TOKEN           API token for authentication
-    GITHUB_REPOSITORY      Repository in owner/repo format
-
-  GitLab:
-    GITLAB_TOKEN           API token (or CI_JOB_TOKEN in GitLab CI)
-    CI_PROJECT_NAMESPACE   Project namespace (auto-set in GitLab CI)
-    CI_PROJECT_NAME        Project name (auto-set in GitLab CI)
-    CI_MERGE_REQUEST_IID   MR number (auto-set in GitLab CI)
-
-  Bitbucket:
-    BITBUCKET_USERNAME     Username for basic auth
-    BITBUCKET_APP_PASSWORD App password for authentication
-    BITBUCKET_WORKSPACE    Workspace name
-    BITBUCKET_REPO_SLUG    Repository slug
-    BITBUCKET_PR_ID        PR number (auto-set in Bitbucket Pipelines)
-
-Examples:
-  # GitHub (auto-detected)
-  coverctl pr-comment --pr 123
-
-  # GitLab (in CI, auto-detects everything)
-  coverctl pr-comment --provider gitlab
-
-  # Bitbucket with explicit values
-  coverctl pr-comment --provider bitbucket --owner myworkspace --repo myrepo --pr 45
-
-  # Dry run to preview comment
-  coverctl pr-comment --pr 123 --dry-run`,
-
-	"mcp": `coverctl mcp - MCP (Model Context Protocol) server for AI agents
-
-Usage:
-  coverctl mcp <subcommand> [flags]
-
-Subcommands:
-  serve       Start the MCP server (STDIO transport)
-
-Flags for 'serve':
-  -c, --config string    Config file path (default ".coverctl.yaml")
-  -p, --profile string   Coverage profile path (default ".cover/coverage.out")
-      --history string   History file path (default ".cover/history.json")
-
-Description:
-  The MCP server enables AI agents (like Claude) to interact with coverctl
-  programmatically. It exposes coverage tools and resources via the Model
-  Context Protocol using STDIO transport.
-
-Tools (actions):
-  check     Run coverage tests and enforce policy thresholds
-  report    Analyze an existing coverage profile
-  record    Record current coverage to history
-
-Resources (read-only queries):
-  coverctl://debt      Coverage debt metrics
-  coverctl://trend     Coverage trends over time
-  coverctl://suggest   Threshold recommendations
-  coverctl://config    Current configuration
-
-Claude Desktop Configuration:
-  Add to ~/.config/claude/claude_desktop_config.json:
-
-  {
-    "mcpServers": {
-      "coverctl": {
-        "command": "coverctl",
-        "args": ["mcp", "serve"],
-        "cwd": "/path/to/your/go/project"
-      }
-    }
-  }
-
-Examples:
-  coverctl mcp serve
-  coverctl mcp serve -c custom.yaml
-  coverctl mcp serve --history .cover/history.json
-  coverctl mcp doctor                  # validate first-run setup
-  coverctl mcp doctor -c custom.yaml   # validate against a non-default config
-
-Subcommands:
-  serve   Start the MCP server (stdio).
-  doctor  Run first-run validation checks. Reports PASS/FAIL with
-          remediation per step: binary on PATH, working-directory
-          markers, config resolvable, MCP server construction, tool
-          dispatch smoke, mode auto-detect. Returns 0 only when every
-          check passes.`,
-
-	"survey": `coverctl survey - Sean Ellis 40% PMF feedback prompt
-
-Asks one question:
-  How would you feel if you could no longer use coverctl?
-
-Responses are appended to ~/.coverctl/survey.jsonl. Nothing is
-transmitted; aggregation is opt-in via the trace donation pipeline
-(deferred per docs/design/gtm-metrics-spec.md).
-
-Usage:
-  coverctl survey                    # interactive prompt
-  coverctl survey --answer very      # scripted: very|somewhat|not|skip
-  coverctl survey --data-dir ./tmp   # override storage location
-
-Why we ask:
-  The Sean Ellis 40% threshold is the standard PMF benchmark. If at
-  least 40% of users would be very disappointed without the product,
-  scaling GTM is justified; below that threshold we go back to
-  discovery before investing in growth.`,
-}
-
-func commandHelp(cmd string, w io.Writer) int {
-	if help, ok := commandHelpText[cmd]; ok {
-		fmt.Fprintln(w, help)
-		return 0
-	}
-	fmt.Fprintf(w, "Unknown command: %s\n\n", cmd)
-	usage(w)
-	return 2
-}
-
-func runCompletion(args []string, stdout, stderr io.Writer) int {
-	if len(args) < 1 {
-		fmt.Fprintln(stderr, "Usage: coverctl completion <bash|zsh|fish>")
-		return 2
-	}
-
-	switch args[0] {
-	case "bash":
-		fmt.Fprintln(stdout, bashCompletion)
-	case "zsh":
-		fmt.Fprintln(stdout, zshCompletion)
-	case "fish":
-		fmt.Fprintln(stdout, fishCompletion)
-	default:
-		fmt.Fprintf(stderr, "Unknown shell: %s\nSupported: bash, zsh, fish\n", args[0])
-		return 2
-	}
-	return 0
-}
-
-const bashCompletion = `# coverctl bash completion
-_coverctl() {
-    local cur prev commands global_flags
-    COMPREPLY=()
-    cur="${COMP_WORDS[COMP_CWORD]}"
-    prev="${COMP_WORDS[COMP_CWORD-1]}"
-    commands="check run watch init detect report badge trend record suggest debt ignore mcp survey help version completion c r w i"
-    global_flags="-q --quiet --no-color --ci --debug"
-
-    if [[ ${COMP_CWORD} -eq 1 ]]; then
-        COMPREPLY=( $(compgen -W "${commands} ${global_flags}" -- ${cur}) )
-        return 0
-    fi
-
-    case "${prev}" in
-        -c|--config)
-            COMPREPLY=( $(compgen -f -X '!*.yaml' -- ${cur}) )
-            return 0
-            ;;
-        -p|--profile)
-            COMPREPLY=( $(compgen -f -X '!*.out' -- ${cur}) )
-            return 0
-            ;;
-        -o|--output)
-            COMPREPLY=( $(compgen -W "text json html" -- ${cur}) )
-            return 0
-            ;;
-        --strategy)
-            COMPREPLY=( $(compgen -W "current aggressive conservative" -- ${cur}) )
-            return 0
-            ;;
-        --style)
-            COMPREPLY=( $(compgen -W "flat flat-square" -- ${cur}) )
-            return 0
-            ;;
-        completion)
-            COMPREPLY=( $(compgen -W "bash zsh fish" -- ${cur}) )
-            return 0
-            ;;
-        mcp)
-            COMPREPLY=( $(compgen -W "serve doctor" -- ${cur}) )
-            return 0
-            ;;
-    esac
-
-    COMPREPLY=( $(compgen -W "-c --config -p --profile -d --domain -o --output -f --force -h --help -q --quiet --no-color --ci --uncovered --diff --merge --show-delta --history --fail-under --ratchet --validate --tags --race --short -v --run --timeout --max-runtime --test-arg" -- ${cur}) )
-}
-complete -F _coverctl coverctl`
-
-const zshCompletion = `#compdef coverctl
-
-_coverctl() {
-    local -a commands
-    commands=(
-        'check:Run coverage and enforce policy'
-        'c:Run coverage and enforce policy (alias)'
-        'run:Run coverage only, produce artifacts'
-        'r:Run coverage only (alias)'
-        'watch:Watch for file changes and re-run coverage'
-        'w:Watch for file changes (alias)'
-        'init:Interactive setup wizard'
-        'i:Interactive setup wizard (alias)'
-        'detect:Autodetect domains and write config'
-        'report:Analyze an existing profile'
-        'badge:Generate an SVG coverage badge'
-        'trend:Show coverage trends over time'
-        'record:Record current coverage to history'
-        'suggest:Suggest optimal coverage thresholds'
-        'debt:Show coverage debt report'
-        'ignore:Show configured excludes and ignore advice'
-        'mcp:MCP server for AI agents'
-        'help:Show help for a command'
-        'version:Show version information'
-        'completion:Generate shell completion scripts'
-    )
-
-    _arguments -C \
-        '-q[Suppress non-essential output]' \
-        '--quiet[Suppress non-essential output]' \
-        '--no-color[Disable colored output]' \
-        '--ci[CI mode: quiet + GitHub Actions annotations]' \
-        '1: :->command' \
-        '*: :->args'
-
-    case $state in
-        command)
-            _describe 'command' commands
-            ;;
-        args)
-            case $words[2] in
-                check|c|run|r|watch|w|report|badge|trend|record|suggest|debt|ignore|init|i|detect)
-                    _arguments \
-                        '-c[Config file path]:file:_files -g "*.yaml"' \
-                        '--config[Config file path]:file:_files -g "*.yaml"' \
-                        '-p[Coverage profile path]:file:_files -g "*.out"' \
-                        '--profile[Coverage profile path]:file:_files -g "*.out"' \
-                        '--from-profile[Use existing coverage profile instead of running tests]' \
-                        '-d[Filter to domain]:domain:' \
-                        '--domain[Filter to domain]:domain:' \
-                        '-o[Output format]:format:(text json html)' \
-                        '--output[Output format]:format:(text json html)' \
-                        '-f[Force overwrite]' \
-                        '--force[Force overwrite]' \
-                        '--uncovered[Show only files with 0% coverage]' \
-                        '--diff[Show coverage for changed files]:ref:' \
-                        '--merge[Merge additional profile]:file:_files -g "*.out"' \
-                        '--show-delta[Show coverage change from previous run]' \
-                        '--history[History file path]:file:_files -g "*.json"' \
-                        '--fail-under[Fail if coverage below threshold]:percent:' \
-                        '--ratchet[Fail if coverage decreases]' \
-                        '--validate[Validate config without running tests]' \
-                        '--tags[Build tags]:tags:' \
-                        '--race[Enable race detector]' \
-                        '--short[Skip long-running tests]' \
-                        '-v[Verbose test output]' \
-                        '--run[Run tests matching pattern]:pattern:' \
-                        '--test-run[Run tests matching pattern]:pattern:' \
-                        '--timeout[Test timeout]:duration:' \
-                        '--test-arg[Additional test argument]:arg:' \
-                        '--language[Override language detection]:lang:(go python nodejs rust java)'
-                    ;;
-                completion)
-                    _arguments '1:shell:(bash zsh fish)'
-                    ;;
-                mcp)
-                    _arguments '1:subcommand:(serve)'
-                    ;;
-            esac
-            ;;
-    esac
-}
-
-_coverctl "$@"`
-
-const fishCompletion = `# coverctl fish completion
-complete -c coverctl -f
-
-# Global flags
-complete -c coverctl -s q -l quiet -d "Suppress non-essential output"
-complete -c coverctl -l no-color -d "Disable colored output"
-complete -c coverctl -l ci -d "CI mode: quiet + GitHub Actions annotations"
-
-# Commands
-complete -c coverctl -n "__fish_use_subcommand" -a "check" -d "Run coverage and enforce policy"
-complete -c coverctl -n "__fish_use_subcommand" -a "c" -d "Run coverage and enforce policy (alias)"
-complete -c coverctl -n "__fish_use_subcommand" -a "run" -d "Run coverage only, produce artifacts"
-complete -c coverctl -n "__fish_use_subcommand" -a "r" -d "Run coverage only (alias)"
-complete -c coverctl -n "__fish_use_subcommand" -a "watch" -d "Watch for file changes and re-run coverage"
-complete -c coverctl -n "__fish_use_subcommand" -a "w" -d "Watch for file changes (alias)"
-complete -c coverctl -n "__fish_use_subcommand" -a "init" -d "Interactive setup wizard"
-complete -c coverctl -n "__fish_use_subcommand" -a "i" -d "Interactive setup wizard (alias)"
-complete -c coverctl -n "__fish_use_subcommand" -a "detect" -d "Autodetect domains and write config"
-complete -c coverctl -n "__fish_use_subcommand" -a "report" -d "Analyze an existing profile"
-complete -c coverctl -n "__fish_use_subcommand" -a "badge" -d "Generate an SVG coverage badge"
-complete -c coverctl -n "__fish_use_subcommand" -a "trend" -d "Show coverage trends over time"
-complete -c coverctl -n "__fish_use_subcommand" -a "record" -d "Record current coverage to history"
-complete -c coverctl -n "__fish_use_subcommand" -a "suggest" -d "Suggest optimal coverage thresholds"
-complete -c coverctl -n "__fish_use_subcommand" -a "debt" -d "Show coverage debt report"
-complete -c coverctl -n "__fish_use_subcommand" -a "ignore" -d "Show configured excludes"
-complete -c coverctl -n "__fish_use_subcommand" -a "mcp" -d "MCP server for AI agents"
-complete -c coverctl -n "__fish_use_subcommand" -a "help" -d "Show help for a command"
-complete -c coverctl -n "__fish_use_subcommand" -a "version" -d "Show version information"
-complete -c coverctl -n "__fish_use_subcommand" -a "completion" -d "Generate shell completion"
-
-# Flags for all commands
-complete -c coverctl -s c -l config -d "Config file path" -r -F
-complete -c coverctl -s p -l profile -d "Coverage profile path" -r -F
-complete -c coverctl -l from-profile -d "Use existing coverage profile instead of running tests"
-complete -c coverctl -s d -l domain -d "Filter to specific domain" -r
-complete -c coverctl -s o -l output -d "Output format" -r -a "text json html"
-complete -c coverctl -s f -l force -d "Force overwrite"
-complete -c coverctl -s h -l help -d "Show help"
-complete -c coverctl -l uncovered -d "Show only files with 0% coverage"
-complete -c coverctl -l diff -d "Show coverage for changed files" -r
-complete -c coverctl -l merge -d "Merge additional coverage profile" -r -F
-complete -c coverctl -l show-delta -d "Show coverage change from previous run"
-complete -c coverctl -l history -d "History file path" -r -F
-complete -c coverctl -l fail-under -d "Fail if coverage below threshold" -r
-complete -c coverctl -l ratchet -d "Fail if coverage decreases"
-complete -c coverctl -l validate -d "Validate config without running tests"
-complete -c coverctl -l tags -d "Build tags (e.g., integration,e2e)" -r
-complete -c coverctl -l race -d "Enable race detector"
-complete -c coverctl -l short -d "Skip long-running tests"
-complete -c coverctl -s v -d "Verbose test output"
-complete -c coverctl -l run -d "Run tests matching pattern" -r
-complete -c coverctl -l test-run -d "Run tests matching pattern" -r
-complete -c coverctl -l timeout -d "Test timeout (e.g., 10m, 1h)" -r
-complete -c coverctl -l test-arg -d "Additional argument passed to go test" -r
-complete -c coverctl -l language -d "Override language detection" -r -a "go python nodejs rust java"
-
-# Completion subcommand
-complete -c coverctl -n "__fish_seen_subcommand_from completion" -a "bash zsh fish"
-
-# MCP subcommand
-complete -c coverctl -n "__fish_seen_subcommand_from mcp" -a "serve" -d "Start the MCP server"`
